// src/enums/dragon-elements.ts
enum DragonElements {
    Terra = "terra",
    Flame = "flame",
    Sea = "sea",
    Nature = "nature",
    Electric = "electric",
    Ice = "ice",
    Metal = "metal",
    Dark = "dark",
    Light = "light",
    War = "war",
    Pure = "pure",
    Legend = "legend",
    Primal = "primal",
    Wind = "wind",
    Time = "time",
    Happy = "happy",
    Chaos = "chaos",
    Magic = "magic",
    Soul = "soul",
    Beauty = "beauty",
    Dream = "dream",
    Physical = "physical",
}

export default DragonElements